修改uchardet为0.07版本，并优化相关解析逻辑

+++ uchardet/CharDistribution.cpp
@@ -55,9 +55,8 @@ float CharDistributionAnalysis::GetConfidence(void)
   if (mTotalChars <= 0 || mFreqChars <= mDataThreshold)
     return SURE_NO;
 
-  if (mTotalChars != mFreqChars)
-  {
-    float r = static_cast<float>(mFreqChars) / (static_cast<float>(mTotalChars - mFreqChars) * mTypicalDistributionRatio);
+  if (mTotalChars != mFreqChars) {
+    float r = mFreqChars / ((mTotalChars - mFreqChars) * mTypicalDistributionRatio);
 
     if (r < SURE_YES)
       return r;
