+++ uchardet/JpCntx.cpp
@@ -186,7 +186,7 @@ float  JapaneseContextAnalysis::GetConfidence(void)
 {
   //This is just one way to calculate confidence. It works well for me.
   if (mTotalRel > mDataThreshold)
-    return static_cast<float>(mTotalRel - mRelSample[0]) / static_cast<float>(mTotalRel);
+    return ((float)(mTotalRel - mRelSample[0]))/mTotalRel;
   else 
     return (float)DONT_KNOW;
 }
@@ -198,13 +198,13 @@ PRInt32 SJISContextAnalysis::GetOrder(const char* str, PRUint32 *charLen)
   if (((unsigned char)*str >= (unsigned char)0x81 && (unsigned char)*str <= (unsigned char)0x9f) ||
       ((unsigned char)*str >= (unsigned char)0xe0 && (unsigned char)*str <= (unsigned char)0xfc))
       *charLen = 2;
-  else 
+  else
       *charLen = 1;
 
   //return its order if it is hiragana
-  if (*str == '\202' && 
-        (unsigned char)*(str+1) >= (unsigned char)0x9f && 
-        (unsigned char)*(str+1) <= (unsigned char)0xf1)
+  if (*str == '\202' &&
+      (unsigned char)*(str+1) >= (unsigned char)0x9f &&
+      (unsigned char)*(str+1) <= (unsigned char)0xf1)
     return (unsigned char)*(str+1) - (unsigned char)0x9f;
   return -1;
 }
@@ -213,17 +213,17 @@ PRInt32 EUCJPContextAnalysis::GetOrder(const char* str, PRUint32 *charLen)
 {
   //find out current char's byte length
   if ((unsigned char)*str == (unsigned char)0x8e ||
-      ((unsigned char)*str >= (unsigned char)0xa1 && 
-      (unsigned char)*str <= (unsigned char)0xfe))
+      ((unsigned char)*str >= (unsigned char)0xa1 &&
+       (unsigned char)*str <= (unsigned char)0xfe))
       *charLen = 2;
   else if ((unsigned char)*str == (unsigned char)0x8f)
-    *charLen = 3;
+      *charLen = 3;
   else
-    *charLen = 1;
+      *charLen = 1;
 
   //return its order if it is hiragana
   if ((unsigned char)*str == (unsigned char)0xa4 &&
-      (unsigned char)*(str+1) >= (unsigned char)0xa1 && 
+      (unsigned char)*(str+1) >= (unsigned char)0xa1 &&
       (unsigned char)*(str+1) <= (unsigned char)0xf3)
      return (unsigned char)*(str+1) - (unsigned char)0xa1;
   return -1;
